feat: force to set the 'direction' flag to the roads if such flag is missed
If the is no 'direction' flag presented in the edge
attributes, we estimate the azimuth (clockwise, zero is north)
from the edge's start and end points.
The direction is quantized to four values (north/east/south/west) for
now, but could be extended to 8 values (including, e.g., north-west,
sourth-east, etc.) if needed.